requirements.txt
setuptools~=50.3.2
pandas~=1.1.4
tensorflow-gpu~=2.3.1